requirements.txt
quart
PyPDF2
python-docx
pandas
chardet
markdown
beautifulsoup4
ebooklib
html2text
langchain
faiss-cpu
sqlalchemy
sentence_transformers
